requirements.txt
ipykernel
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
joblib
streamlit
pillow